fix(agent-runtime): satisfy CI readability and lint gates
