+++ .ci/readability-baseline.env
@@ -1,11 +1,11 @@
 # Generated by scripts/readability-ratchet.sh
-PROD_RS_TOTAL=594
-PROD_FILES_GT300=234
-PROD_FILES_GT500=86
+PROD_RS_TOTAL=523
+PROD_FILES_GT300=209
+PROD_FILES_GT500=84
 PROD_FILES_GT1000=22
-PROD_MAX_FILE_LINES=2727
+PROD_MAX_FILE_LINES=2829
 PROD_MAX_FILE_PATH=crates/temper-server/src/storage/mod.rs
 ALLOW_CLIPPY_COUNT=35
-ALLOW_DEAD_CODE_COUNT=19
+ALLOW_DEAD_CODE_COUNT=16
 PROD_PRINTLN_COUNT=247
-PROD_UNWRAP_CI_OK_COUNT=126
+PROD_UNWRAP_CI_OK_COUNT=132
